requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.13.1+cpu
torchvision==0.14.1+cpu
pytorch-lightning>=1.9.0,<2.0.0
numpy>=1.21.0
matplotlib>=3.4.0
tqdm>=4.62.0
wandb>=0.12.0  # for experiment tracking
hydra-core>=1.3.0  # for config management
omegaconf>=2.3.0  # for config management
pillow>=8.3.0
scikit-learn>=0.24.0
seaborn>=0.11.0  # for plotting 